# account/openapi.yaml
openapi: 3.0.1
info:
  title: Account API
  description: API for managing accounts
  version: 1.0.0

paths:
  /accounts:
    get:
      summary: Get all accounts
      operationId: getAccounts
      tags:
        - Accounts
      responses:
        '200':
          description: Successfully retrieved list of accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Account'
    post:
      summary: Create a new account
      operationId: createAccount
      tags:
        - Accounts
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Account'
      responses:
        '201':
          description: Account created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'

  /accounts/{id}:
    get:
      summary: Get account by ID
      operationId: getAccountById
      tags:
        - Accounts
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: long
      responses:
        '200':
          description: Successfully retrieved account
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        '404':
          description: Account not found
    put:
      summary: Update an account
      operationId: updateAccount
      tags:
        - Accounts
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: long
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Account'
      responses:
        '200':
          description: Account updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        '404':
          description: Account not found
    delete:
      summary: Delete an account
      operationId: deleteAccount
      tags:
        - Accounts
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: long
      responses:
        '204':
          description: Account deleted successfully
        '404':
          description: Account not found

components:
  schemas:
    Account:
      type: object
      properties:
        id:
          type: long
          example: 1
        name:
          type: string
          example: John Doe
        email:
          type: string
          example: john.doe@example.com
        accountNumber:
          type: string
          example: "1234567890"
        accountType:
          type: string
          example: "Savings"
        dob:
          type: string
          format: date
          example: "1990-01-01"
        bankCode:
          type: string
          example: "XYZ123"